% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\docType{data}
\name{attributes}
\alias{attributes}
\title{Attributes from input file}
\format{
List with 7 elements
\describe{
\item{analysis_type}{String with analysis type to filter stations}
\item{main_url}{Main URL of the MS2Soft platform by State}
\item{dl_path}{Downloads folder of the system}
\item{dot}{Texas Department of Transportation Acronym}
\item{offset}{Offset value to conform URL}
\item{a}{Value "a" to conform URL}
\item{class_table}{Table of class IDs and reclassification}
}
}
\usage{
attributes
}
\description{
List containing attributes for the MS2Soft data download
}
\examples{
attributes
}
\keyword{datasets}
