% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_points.R
\docType{data}
\name{base_points}
\alias{base_points}
\title{Example base waypoints for TT Methods}
\format{
Dataframe with 24 observations of 5 variables
\describe{
\item{Name}{Name of the given waypoint}
\item{Description}{Description field used by Google Earth (not used)}
\item{Route}{Route identifier}
\item{Point}{Point number within route}
\item{geometry}{Special Features geometry column}
}
}
\usage{
base_points
}
\description{
Dataframe containing waypoints for two routes (one by direction) of an
example route to gather travel times
}
\examples{
base_points
}
\keyword{datasets}
