% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_all_data.R
\name{ms2_all_data}
\alias{ms2_all_data}
\title{Get traffic data for all stations in the stations list file}
\usage{
ms2_all_data(
  dates_list,
  main_url,
  analysis_type = c("class", "perm", "short", "speed"),
  stations_list,
  attributes
)
}
\arguments{
\item{dates_list}{The dates list obtained with the \code{ms2_all_data} function}

\item{main_url}{The main URL, can be omitted if declaring the attribute table}

\item{analysis_type}{The type of analysis to be performed, between
class, perm and short}

\item{stations_list}{The dataframe containing the stations to be consulted.}

\item{attributes}{Attributes list obtained by the \code{ms2_attributes} function}
}
\value{
Downloaded files within the created stations folder
}
\description{
Get traffic data for all stations in the stations list file, downloads data and stores files in a
\verb{stations/} folder within the working directory.
Downloads traffic data of the types  permanent, short term or classification
}
