% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_all_dates.R
\name{ms2_all_dates}
\alias{ms2_all_dates}
\title{Get a list of available dates for a set of MS2Soft Stations}
\usage{
ms2_all_dates(
  main_url,
  analysis_type = c("class", "perm", "short", "speed"),
  offset,
  a,
  quiet = FALSE,
  day_type = "all",
  sampling = FALSE,
  sample_weeks = 1,
  stations_list,
  attributes,
  gather_dates = FALSE
)
}
\arguments{
\item{main_url}{The URL of the state MS2Soft platform}

\item{analysis_type}{The type of analysis to be performed, between
class, perm and short}

\item{offset}{The offset value within the MS2Soft platform URL}

\item{a}{The "a" value within the MS2Soft platform URL}

\item{quiet}{Defaults to TRUE, wether to show the downloading progress}

\item{day_type}{Defines wether only weekdays or weekends are selected. Defaults to all}

\item{sampling}{Generates a sample of the downloaded data, useful when
having too many dates}

\item{sample_weeks}{Number of weeks to sample from each year}

\item{stations_list}{The dataframe containing the stations to be consulted,
avoids entering the \code{start_date} and \code{end_date} and allows for different
values across stations. Useful when running multiple stations.}

\item{attributes}{Attributes list obtained by the \code{ms2_attributes} function}

\item{gather_dates}{Input TRUE to gather dates from previously downloaded
date vectors in the TEMP folder}
}
\value{
A list of dates when data is available
}
\description{
Download a list of dates in which data is available for the analysis type
selected
}
