% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_attributes.R
\name{ms2_attributes}
\alias{ms2_attributes}
\title{Import MS2Soft attributes as list}
\usage{
ms2_attributes(path = "inputs/input_file.xlsx")
}
\arguments{
\item{path}{path to the attributes list file and class table, usually left
as default if using recomended folder structure}
}
\value{
A list containing the attributes required for the workflow
}
\description{
Imports the attributes  list from an input file containing the metadata required
to download data
}
