% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_createdb.R
\name{ms2_createdb}
\alias{ms2_createdb}
\title{Create a database from files downloaded}
\usage{
ms2_createdb(
  analysis_type = NULL,
  attributes = attributes,
  class_number = 1,
  get_interval = FALSE
)
}
\arguments{
\item{analysis_type}{The type of analysis to be performed, between
class, perm and short}

\item{attributes}{Attributes list obtained by the \code{ms2_attributes()} function}

\item{class_number}{Column number for the reclassification table to name
columns}

\item{get_interval}{Defaults to FALSE, extracts total data. If TRUE, extracts
data by interval}
}
\value{
A dataframe containing all traffic data stored in the  \verb{stations/}
folder
}
\description{
Imports data stored in the stations folder and arranges it into a dataframe
}
