% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_data.R
\name{ms2_data}
\alias{ms2_data}
\title{Get traffic data for a given station}
\usage{
ms2_data(
  station,
  date,
  main_url,
  analysis_type = c("class", "perm", "short", "speed"),
  stations_list,
  attributes
)
}
\arguments{
\item{station}{The station ID for which the data will be requested}

\item{date}{The date in DD/MM/YYY format as read by the URL}

\item{main_url}{The main URL, can be omitted if declaring the attribute table}

\item{analysis_type}{The type of analysis to be performed, between
class, perm and short}

\item{stations_list}{The dataframe containing the stations to be consulted.}

\item{attributes}{Attributes list obtained by the \code{ms2_attributes} function}
}
\value{
Downloaded files within the created stations folder
}
\description{
Downloads data of the types for analysis types permanent, short term or classification
and stores files in a \verb{stations/} folder within the working directory.
}
