% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_dates.R
\name{ms2_dates}
\alias{ms2_dates}
\title{Get a vector of available dates for an MS2Soft Station}
\usage{
ms2_dates(
  station,
  main_url,
  analysis_type = c("class", "perm", "short", "speed"),
  offset,
  a,
  start_date,
  end_date,
  quiet = TRUE,
  day_type = c("weekday", "weekend", "all"),
  sampling = TRUE,
  sample_weeks = 1,
  stations_list,
  attributes
)
}
\arguments{
\item{station}{The station ID for which the data will be requested}

\item{main_url}{The URL of the state MS2Soft platform}

\item{analysis_type}{The type of analysis to be performed, between
class, perm and short}

\item{offset}{The offset value within the MS2Soft platform URL}

\item{a}{The "a" value within the MS2Soft platform URL}

\item{start_date}{The starting year for the request}

\item{end_date}{The final year to be requested}

\item{quiet}{Defaults to TRUE, wether to show the downloading progress}

\item{day_type}{Defines wether only weekdays or weekends are selected}

\item{sampling}{Generates a sample of the downloaded data, useful when
having too many dates}

\item{sample_weeks}{Number of weeks to sample from each year}

\item{stations_list}{The dataframe containing the stations to be consulted,
avoids entering the \code{start_date} and \code{end_date} and allows for different
values across stations. Useful when running multiple stations.}

\item{attributes}{Attributes list obtained by the \code{ms2_attributes} function}
}
\value{
A vector of dates when data is available
}
\description{
Creates a vector containing all available dates within the time range given
between \code{start_date} and \code{end_date} in years
}
