% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_inputs.R
\name{ms2_inputs}
\alias{ms2_inputs}
\title{Create  MS2Soft attributes and class input file}
\usage{
ms2_inputs(
  path = "inputs/input_file.xlsx",
  fill = FALSE,
  state = NULL,
  analysis_type = NULL,
  start_year = NULL,
  end_year = NULL,
  downloads_folder = NULL,
  dot = NULL,
  main_url = NULL,
  offset = NULL,
  a = NULL,
  class_ids,
  reclass
)
}
\arguments{
\item{path}{path to the stations list file, usually left as default if using
recomended folder structure}

\item{fill}{Defaults to FALSE, to produce an empty input file}

\item{state}{The US State from where to download data}

\item{analysis_type}{The analysis type between \code{perm}, \code{class} and
\code{short}}

\item{start_year}{The starting year for the analysis}

\item{end_year}{The endind year for the analysis}

\item{downloads_folder}{The path to the downloads folder}

\item{dot}{The department of transportation acronym corresponding to the
platform being consulted}

\item{main_url}{The core URL used to host the main page of the platform}

\item{offset}{The offset parameter from the URL when consulted as frames}

\item{a}{The a parameter from the URL when consulted as frames}

\item{class_ids}{The classifications IDs given to classified traffic data}

\item{reclass}{Reclassification names, use \code{class_ids} if no reclassification
is needed}
}
\value{
An excel file containing the columns required for the analysis
}
\description{
Creates an input file with attributes and class equivalencies table.
Attributes and class table can be defined as arguments or the file can be
made empty to fill out later
}
