% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_sample_dates.R
\name{ms2_sample_dates}
\alias{ms2_sample_dates}
\title{Sample a dates list}
\usage{
ms2_sample_dates(dates_list, sample_weeks = 1)
}
\arguments{
\item{dates_list}{a dates list obtained with the \strong{ms2_all_dates} function}

\item{sample_weeks}{the amount of weeks to sample each month}
}
\value{
A reduced dates list with x amounts of weeks sampled by month
}
\description{
Get a reduced dates list by sampling by month and year, only selects weekdays
}
