% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_stationsCols.R
\name{ms2_stationsCols}
\alias{ms2_stationsCols}
\title{Create  MS2Soft stations list basefile}
\usage{
ms2_stationsCols(path = "inputs/stations_list.xlsx")
}
\arguments{
\item{path}{path to the stations list file, usually left as default if using
recomended folder structure}
}
\value{
An excel file containing the columns required for the analysis
}
\description{
Creates a stations list that can be used to fill with stations metadata
}
