% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_times.R
\docType{data}
\name{travel_times}
\alias{travel_times}
\title{Example of travel times query results}
\format{
Dataframe with 506 observations of 18 variables
\describe{
\item{alternative_id}{}
\item{leg_id}{}
\item{summary}{}
\item{distance_m}{}
\item{distance_text}{}
\item{duration_s}{}
\item{duration_text}{}
\item{duration_in_traffic_s}{}
\item{duration_in_traffic_text}{}
\item{departure_time}{}
\item{arrival_time}{}
\item{TOD}{}
\item{route}{}
\item{Code}{}
\item{geometry}{}
\item{travel_time}{}
\item{distance}{}
\item{speed}{}
}
}
\usage{
travel_times
}
\description{
Dataframe containing resulting queries from travel times example
}
\examples{
travel_times
}
\keyword{datasets}
