% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_export.R
\name{tt_export}
\alias{tt_export}
\title{Export Google Travel Results}
\usage{
tt_export(x, file_name, type)
}
\arguments{
\item{x}{Dataframe obtained from \code{tt_travel_time()} or list of matrices obtained
with \code{tt_matrix}}

\item{file_name}{Full file name including path and extension}

\item{type}{Type of data for the exported file, either "matrix", "dataframe"
or "shp"}
}
\description{
Export results obtained from travel time requests as dataframe or
shapefile obtained with \code{tt_travel_time()}. Or travel time, speed or distance
matrices obtained with \code{tt_matric()}.
}
