% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_gather_points.R
\name{tt_gather_points}
\alias{tt_gather_points}
\title{Imports points from spatial data}
\usage{
tt_gather_points(points_path, details_path, sheet = "PointDetails")
}
\arguments{
\item{points_path}{System path containing the points spatial data}

\item{details_path}{System path directing to the file containing the points
details or descriptions}

\item{sheet}{Sheet within the details file containing the point details,
defaults to "PointDetails"}
}
\value{
A special features object with the points for all routes that will
be queried
}
\description{
Used to import points in a geographic format. Requires an excel file containing
a sheet with point details, including a joining code between the spatial file
and the details.
}
