% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_hours_vector.R
\name{tt_hours_vector}
\alias{tt_hours_vector}
\title{Create a vector of hours composing the Time of Day periods}
\usage{
tt_hours_vector(date, hours_numbers)
}
\arguments{
\item{date}{Date to be queried. Must be a future day, at least one day after
the current date}

\item{hours_numbers}{Vector of numeric values of hours}
}
\value{
A vector of hours of type POSIXct
}
\description{
Create a vector of hours composing the Time of Day periods
}
\examples{
hours <- tt_hours_vector(Sys.Date()+1, c(2, 4, 8, 12, 14))
}
